Split TlsConnection::SetEncryptionSecrets in two
https://boringssl-review.googlesource.com/c/boringssl/+/40127 changes how
encryption secrets are passed from BoringSSL to QUIC. Instead of setting
both the read and write secrets in a single call, they will be set in two
separate calls which will occur at different times.

This is the first CL to prepare for that change. It splits
SetEncryptionSecrets into SetWriteSecret and SetReadSecret. This CL does
not introduce a behavior change: SetWriteSecret and SetReadSecret are
called at the same time SetEncryptionSecrets formerly was called.

gfe-relnote: Split method in two, no behavior change
PiperOrigin-RevId: 297417937
Change-Id: I857505c93111ee0c1d4d2685d085630581feb25d