Have ParsedQuicVersionIsValid also validate transport versions
As part of investigating <https://crbug.com/1061509>,  we realized that since versions can be saved to disk, it's possible to try to create an invalid ParsedQuicVersion based on data from a previous build before a given version was deprecated. So to better enforce our assumptions, this CL changes ParsedQuicVersionIsValid to also check whether the transport version is known by the current codebase.

gfe-relnote: change a DCHECK, no changes in production
PiperOrigin-RevId: 301259614
Change-Id: I11051d88ddc7f9eae426858902d2b82355578910