Change QuartcPeerTest to use a MockSendAlgorithm.
This removes the dependency on behavior of BBR(v2), so the test now covers only
how QuartcPeer handles whatever bandwidth estimate it receives.

The test uses a fixed bandwidth estimate in order to keep everything simple.
The estimate is equal to the simulator's link bandwidth.

This should remove flakiness and brittleness.  This test has historically failed
whenever unrelated changes in the default send algorithm alter the behavior
observed in the test.  Now it should be more resilient to such unrelated
changes.

gfe-relnote: n/a (Quartc test only)
PiperOrigin-RevId: 301905253
Change-Id: I8d0db3657e930002d3b23eb1300cfbbe1e212c9b