Do not send a MAX_PUSH_ID frame if SetMaxPushId() has not been called yet.
This is a correctness issue.  A MAX_PUSH_ID frame with value 0 signals to
the server that it is allowed to create a single push stream, with push ID 0.  Not sending a MAX_PUSH_ID frame signals to the server that it is not yet allowed to create any push streams.

gfe-relnote: Client-side behavior change only.  Not protected.
PiperOrigin-RevId: 303775633
Change-Id: Iae4f4984ae783e6a95137b3a701d99ccf81e482e