gfe-relnote: Change QuicRstStream::ietf_error_code to 64 bits and move it out of the union with error_code.  Behavior change in IETF QUIC only, protected by gfe2_reloadable_flag_quic_enable_version_draft_25_v3 and gfe2_reloadable_flag_quic_enable_version_draft_27.
Multiple tests rely on ietf_error_code being set to the value of error_code.
However, if they stay in a union, and ietf_error_code is changed to 64 bits,
then the upper bits will be unitialized, because error_code has fewer bits.
Fortunately I was planning to move them out of the union so that upper layers
can still interact with QuicRstStreamErrorCode regardless of QUIC version.

PiperOrigin-RevId: 305870819
Change-Id: Iae47ede459094ade6f9dc7fddfc9acadb82fa417