Add support for signing to the CertificateView API.
The signing here is needed to provide a reasonable ProofSource implementation without depending on Chromium or google3, that can be used both in unit tests and by public API users.

The verify here is required to support QUIC crypto in unit tests.

gfe-relnote: n/a (code not used in production)
PiperOrigin-RevId: 307097439
Change-Id: I2d544c5e6cd4b60ed0965dfb93b9476d75524394