Add move assignment operator to QuicStreamSequencer
This is required for upcoming QUIC+TLS ALPN extraction.
A new extractor class will be a member to
QuicBufferedPacketStore::BufferedPacketList which needs
a move assignment operator, and it uses QuicStreamSequencer,
therefore we need that to support move assignment. Similarly,
that requires adding move asignment to QuicStreamSequencerBuffer
which also requires making two const members non-const.

gfe-relnote: n/a, added unused operator
PiperOrigin-RevId: 307459373
Change-Id: I3091cb7a6a27a6b01e528417500c18746f3adc3e