Automated g4 rollback of changelist 307677297.
*** Reason for rollback ***

Some tests have a comment stating that they do not work on Chrome with TLS.  I am not able to reproduce that issue locally on Linux, but it might be the case on other platforms.

*** Original change description ***

Enable TLS for some tests in quic/core/http/end_to_end_tests.cc.

gfe-relnote: n/a, test-only change.

***

PiperOrigin-RevId: 307839934
Change-Id: If05110cc366e7976d086abafdf8b59c0bc8ff48c