gfe-relnote: Add B2LO connection option to ignore inflight_lo in BBRv2.  Protected by gfe2_reloadable_flag_quic_bbr2_ignore_inflight_lo.
This may not be the ideal long term option, but given bandwidth_lo is an existing shorter_term adjustment to loss, I fear inflight_lo is causing excess harm and complexity.  This could overlap with quic_bbr2_avoid_too_low_probe_bw_cwnd, but may be a preferable alternative.

BBRv2 tests pass with this default enabled as well.

PiperOrigin-RevId: 308673519
Change-Id: I6ce8b13dda3ca427919820dab04d135e3e4c8867