Remove QUIC silent close negotiation
The ability to negotiate silent close was never used and is not defined in the IETF QUIC specification. This CL removes this negotiation and makes us always use silent close once the handshake is complete. During the handshake we will continue to send connection close packets.

gfe-relnote: n/a, remove unused code
PiperOrigin-RevId: 309320830
Change-Id: Id3ca8e408a46533d434bbb1b69656e0331f9b16f