Further clean up QuicConfig idle timeout API
Now that all of our clients use the single-parameter API, remove any API that allows setting different values or that uses seconds instead of QuicTime::Delta. Since the idle timeout is not used before the handshake, and required to be negotiated during the handshake, we're guaranteed that its default value is never used and only the maximum value matters. Therefore this CL does not change any behavior.

gfe-relnote: refactor QuicConfig, not flag protected
PiperOrigin-RevId: 309474946
Change-Id: I070efb307b360d0e30d4481dfb3fe24bbec62306