Remove QuicNegotiableValue
This CL reimplements QUIC idle timeout negotiation without QuicNegotiableValue, and removes QuicNegotiableValue. QuicNegotiableValue is considered overly complex and hard to understand and use correctly, and has caused bugs in the past. Idle timeout was the only user of it, so reusing QuicFixedUint32 and bespoke handling was the simplest option. The wire format for idle timeout is unchanged by this CL: it's still encoded as a 32bit integer in seconds for QUIC_CRYPTO and a 62bit integer in milliseconds for QUIC+TLS. Therefore this CL does not change any behavior.

gfe-relnote: refactor idle timeout, no behavior change, not flag protected

Startblock:
  after 2020-05-05 09:00 in US/Pacific
PiperOrigin-RevId: 309988666
Change-Id: I67ae7caad8ec73c05e27c8e49d940a2393d4b2df