Plumb active_connection_id_limit to QuicConfig
We currently do not use active_connection_id_limit anywhere because we do not yet support IETF QUIC connection migration. This CL plumbs that value to QuicConfig to allow us to remember it for 0-RTT and validate that values are not lowered by the server.

This CL also changes the default and minimum values for active_connection_id_limit from 0 to 2, due to recent changes to the spec. This does not change our behavior because we currently never send this transport parameter, nor act in any way on the value received. (This is because we do not send a transport parameter when it is configured to a value that is equal to the default.)

Plumb unused active_connection_id_limit, no behavior change

PiperOrigin-RevId: 310382355
Change-Id: Ifdda3ec3473f1472256788c87ca172fe037a12cb