Re-enable TLS for EndToEndTest.AckNotifierWithPacketLossAndBlockedSocket.
This was an issue with QPACK, not TLS, but since currently every QUIC version
using QPACK also uses TLS, this test being disabled for TLS hid the underlying
issue.

This was caused by cl/309141996, which enabled the QPACK dynamic table for the
test client and server.  This CL fixes the test by enabling the dynamic table
for the local QpackEncoder instance that is solely used to predict the size of
the compressed headers.

Also fix the obsolete class comment for TestAckListener.

Test-only change.

PiperOrigin-RevId: 311118418
Change-Id: I4eb6b29e9f0a3ed8864babdb4200b54b85c8b17b