Allow --host and url to refer to different hosts in quic_client_interop_test_bin
Occasionally it's useful to connect to a server with one hostname but use a
different hostname for the rest of the connection. This allows the host
specified by --host to determine what IP address to connect to, and the host
in the URL to determine what gets sent in SNI, :authority, and anywhere else
a hostname is used. If one of those two is missing, it will default to the
other.

Test tools only

PiperOrigin-RevId: 311444304
Change-Id: I7092e4031031dd41a816bf7befe0e67c5e35645d