Refactor QuicConnection::AddIncomingConnectionId
In order to prepare for draft-28, this CL refactors how we set the original_destination_connection_id on QuicConnection. The original plan for AddIncomingConnectionId was to be used in more places, but today it is only used for the original connection ID, so we're refactoring it to allow saving off that connection ID in a subsequent change. This CL has no behavior change.

Refactor, no behavior change, not flag protected

PiperOrigin-RevId: 313465625