Refactor TlsHandshaker interactions with the session's config
Before this CL, both TlsClientHandshaker and TlsServerHandshaker used to access session->config() directly to read and write transport parameters. This CL refactors this to go through the handshaker delegate, which is implemented by the session. This goes towards our goal of removing the session from the handshakers, but also is needed to allow the session to make some modifications to the transport parameters - that will happen in a subsequent CL aimed at supporting the draft-28 transport parameters.

Refactor, no behavior change, not flag protected

PiperOrigin-RevId: 313495203
Change-Id: I2d8acf8887c40de6be284395a57948dbc1cadc66