Add ipv6_only to QuicUdpSocketApi::Create.
This is necessary for iOS compatibility - setting IP_PKTINFO on an IPv6 socket will always fail.  ipv6_only will cause IP_PKTINFO to be skipped for IPv6 sockets.

Testing is also expanded to cover forcing IPv6 on all platforms, as well as to ensure that ipv6_only does not cause breakages on other platforms.

Add ipv6_only for quicudpsocketapi::create.  this does not change default behavior.

PiperOrigin-RevId: 313669717
Change-Id: I0e33b5ff94ceb299e69a50976062f04b4a3452bb