Avoid crash in QuicDispatcher
Before this CL, when the dispatcher's call to CreateQuicSession returned nullptr, we would crash. This CL changes that to instead log a QUIC_BUG and return. This makes debugging QuicDispatcher tests a lot easier because they now log a test failure instead of crashing and disabling test results.

No behavior change, crash not ever seen in production now logs a  gfe_bug instead

PiperOrigin-RevId: 317697386
Change-Id: I21245840870495ac0b4b7b3cef27f10e27ba7cd3