Make sure ParseQuicTagVector returns an empty array on an empty string
This CL makes ParseQuicTagVector match Chrome's net::ParseQuicConnectionOptions so we can remove that and use this instead. The only functional difference is that an empty string needs to return an empty vector. This function is currently only used in our test client so the change does not require flag protection.

Test-only change

PiperOrigin-RevId: 317945520
Change-Id: I29a2bec31ba39f37ff00f7352c288f9dcd51ea72