Fix memory safety issues in e2e test
This test contained multiple memory safety issues that would trigger if something failed. For example, if the handshake fails, there might not be a server connection and session created, and our code would previously use uninitialized memory as its session. Similarly, we would dereference header values even if we failed to find the right header, leading to use of uninitialized memory.

To fix these issues, this CL refactors some code to convenience methods that perform proper checks, and make sure the test is marked as failed if any of the convenience methods return early due to a missing object.

Note that we need to pause the server thread in order to safely access server variables, and we need to then resume the server thread even if there were failures, therefore tests cannot use ASSERTs when the server thread is paused.

gf-relnote: n/a, test-only change
PiperOrigin-RevId: 319255234
Change-Id: Ic1281bb2dd1217e1303fca1f20cb4275a343cc1c