Remove SSL methods from non-SSL base class
For historical and structural reasons, it has been convenient to have several methods which only apply to crypto protocols (SetSNIServerName, GetSSLCertificateCommonName, and IsSessionResumed) present as pure virtual methods in the base SimpleClient interface.  Implementations for which these methods have no meaning simply stub them out.

This is gross, and an impediment to some refactoring I am doing.

So, this CL rectifies the situation.  It removes the base-class methods, and implements these three operations as free functions which use RAII to implement their functionality for TLS and QUIC separately.

Functionally, this CL is a no-op.

N/a test changes only

PiperOrigin-RevId: 319982508
Change-Id: I70762778428f1a826f1d72e8ace77e0c7f9880f1