In quicpacketcreator::serializepacket, set encrypted_buffer.buffer to nullptr before function returns. no behavior change, not protected.
The avoid_leak_writer_buffer flags has been default enabled but can't be deprecated due to a Chrome OS test failure in b/157989434. I suspect in that test environment, the std::move-d release_buffer(a std::function) is not set to an empty state, so the destructor of encrypted_buffer is still trying free the encrypted buffer. This is not observed in other environments.

PiperOrigin-RevId: 320427430