Update goaway frame field name and type. not protected.
Motivation for type change: remote endpoints might send GOAWAY frames with ID
larger than what could be represented on QuicStreamId (uint32_t).  It is
important that a value like 2^32+8 is not treated as 8 but as the largest
possible QuicStreamId.  The ideal place to handle this is
QuicSpdySession::OnHttp3GoAway(), therefore this CL plumbs 64 bits there from
HttpDecoder::FinishParsing().  For the time being this only means moving the
explicit cast, so there is no behavior change.

There is no need to change SendHttp3GoAway() to handle 64 bit values, because
QuicStreamId is only 32 bits internally, so 2^32-4 is sufficient to send when
doing graceful shutdown (which is not yet implemented).

Motivation for name change: https://github.com/quicwg/base-drafts/pull/3129
introduces GOAWAY frames sent from client to server, in which case the payload
carries a push ID, not a stream ID.

PiperOrigin-RevId: 322393259
Change-Id: I472ec669259d675705790ee12c9404ca3d54195c