Refactor QuicConnection::UpdatePacketContent() to take a different enum to reflect frame type. No behavior change, not protected.
Currently it takes PacketContent which is also used as current_packet_content_ connection state. This state is only needed in gQUIC. However UpdatePacketContent() can be used to update iQUIC connection state to indicate whether current packet is a probing packet or a non-probing packet depending on frames in it.

Adding calls to UpdatePacketContent() in all IETF frame callbacks as well. These calls will be no-op in UpdatePacketContent().

PiperOrigin-RevId: 324700454
Change-Id: I96d35d990ee1727e2f781e9ae11e94113a332ec2