(n/a, not in use) Make IETF probing packets not always contribute to RTT measurement.
Probing packets can be sent on current path and alternative path. It is ideal that when it's sent on alternative path, its ACK doesn't contribute to the RTT measurement on the current path.

This change added a new state for SentPacketState for that purpose, and make SentPacketManager::OnPacketSent() take an extra parameter to indicate if the packet is sent on alternative path.

PiperOrigin-RevId: 328330767
Change-Id: Ibba53390097cfa8cd08362dca3595c13154a6db9