Test-only change
Test that IETF application error codes translated to QuicRstStreamErrorCode and
then back yield the original value.  This is important because upon receiving a
STOP_SENDING frame we send a RESET_STREAM frame (QuicStream::OnStopSending()
calls QuicSession::SendRstStream() passing around a QuicRstStreamErrorCode), and
this guarantees that the sent wire error code will be identical to the one
received.

PiperOrigin-RevId: 328560090
Change-Id: Ifc0d74e2d53fb5a1bcc7f90ce02bac1fad3f0155