Refactor GOAWAY-related accessors.
In Google QUIC, there is a transport layer GOAWAY frame.  IETF QUIC does not
have such a frame, because it is defined in the application layer HTTP/3.

This method prepends a tranport_ prefix to QuicSession's accessors to reflect that they operate on the transport layer GOAWAY frame only.  It also replaces QuicSpdySession's accessors with http3_ prefix with generic ones that access information on the appropriate kind of frame determined by the protocol version.

Refactor with no functional change

PiperOrigin-RevId: 328719322
Change-Id: I028d8e18117fd05b4800beea8181c69a61c239b9