Add method to send HTTP/3 shut down notice GOAWAY frame
Add QuicSpdySession::SendHttp3Shutdown() to send a GOAWAY frame with the
maximum stream ID. This can be used to notify clients that they should stop
creating new streams on the connection. This method is unused.

The behavioral change to QuicSpdySession::SendHttp3GoAway() is safe because it
only appears when callers call SendHttp3Shutdown(), and that method is new and
unused.

PiperOrigin-RevId: 329797508
Change-Id: I8602a3ac56703a7ac0ef2c133e8c8dd88cc3217d