Fix stream ID sent in GOAWAY frame.
The stream ID sent in a HTTP/3 GOAWAY frame is the lowest that can be retried by
the client.  (This is unlike HTTP/2 where the stream ID is the highest that
cannot be retried.)  Fix QuicSpdySession::SendHttp3GoAway() to send the next
stream ID after the largest received one.

Also do not send GOAWAY frame if it could not convey new information to the
client with respect to the previously sent GOAWAY frame.

Protected by FLAGS_quic_reloadable_flag_quic_fix_http3_goaway_stream_id.

PiperOrigin-RevId: 330840957
Change-Id: I3f2a1a18e5d0fb294d39f21c315b8af7ec1a4811