Self-instrument for QUIC library via setsockopt for on-host-monitoring.
go/secure-protocols-on-host-detection is our design to implement
on-host monitoring for connection tracking. It requires code creating
sockets to tag the socket through an additional syscall. Most of our
traffic is covered through Stubby, but Quic also generates a lot of
traffic and manages raw sockets itself.

This CL should close the monitoring gap for Quic.

Given that we already do setsockopt tagging inside Stubby2 and GRPC (which the GFE uses), we chose not to flag protect this code path, as this is basically a tested feature at this point. Please see CL Discussion.

gfe-relnote: Self-instrument for QUIC library via setsockopt for on-host-monitoring.
PiperOrigin-RevId: 330932821
Change-Id: Ibf83cbdbe6792119b83e702e0f702b22bcb6253d