Remove BoundedHuffmanSize().
Benchmarks show that it is almost twice as fast as ExactHuffmanSize() for an
input entirely made up of characters that are encoded on 30 bits each, but takes
about five times as long for an input entirely made up of 'a'.  Typical input is
expected to contain characters with short codes (that's how the Huffman table
was constructed), so ExactHuffmanSize() is a better choice on average.

Also rename ExactHuffmanSize() to HuffmanSize().

Also remove obsolete TODO about binary literals.

PiperOrigin-RevId: 335412149
Change-Id: I526e04417298468f5b83f3f26dc40ab62ecb1960