Switch ParsedQuicVersion::ReservedForNegotiation() to TLS
When the ReservedForNegotiation version was first added, our QUIC+TLS  implementation wasn't as developed as it is today, so we made it a QUIC_CRYPTO version closely resembling Q099. We've now removed support for Q099, and it's more future-proof to have the ReservedForNegotiation version be as close as possible to our leading version, which is h3-29 which uses TLS. This CL also gives the ReservedForNegotiation version its own initial obfuscator salt and retry integrity key and nonce, to ensure that it can't be accidentally decrypted by other versions.

This CL is not flag protected because the ReservedForNegotiation version is only used in tests and the interop tool.

PiperOrigin-RevId: 335701278
Change-Id: I0b3596c665ad843e5a87301e546e2bd472c39aab