Fix NetSLO QUIC prober tests
These broke when we landed cl/335846456 because the tests were hand-crafting QUIC packets. This CL adds a new test-only API to QUIC to perform this parsing, and then uses that API from the NetSLO tests. This CL also slightly tweaks the NetSLO code to populate the connection ID length to improve memory safety.

PiperOrigin-RevId: 335974800
Change-Id: Ide0a6e30b6510f606fd40da91f9b61598883d926