Add unused interface QuicConnection::SendPathChallenge().
This interface is implemented by 2 code paths:
If the PATH_CHALLENGE is sent on the default socket, add the frame as other frames to the current serialize packet in packet creator and flush at its due time. Since the destination address might be different for that packet, use a ScopedPeerAddressContext to flush it if the peer address is different from the one in used at the call site.
If the PATH_CHALLENGE is sent on an alternative socket, send it in a stand alone code path by serializing it into a individual packet and writing it to the wire directly.

PiperOrigin-RevId: 337176402
Change-Id: I90c4b88809209dd59a31a98c58d0de3ecd9c4889