Move protocol flags out of platform so that they are included in QUICHE.
Includes Chrome values updated at https://crrev.com/c/2488754.

Change quic_send_buffer_max_data_slice_size type to uint64.  It was int32
internally and uint32 in Chromium.  Internal infrastructure does not support
DECLARE_uint32.  Changing it to int32 would trip a couple of Chromium tests with
integer literal comparisons that are signedness-sensitive, and it feels wrong to
change them to signed for a flag that is used to set a QuicByteCount type, which
is uint64_t.

PiperOrigin-RevId: 338769159
Change-Id: I1a6743cbe381a640d3c9da6ffb6ece5c64a5879e